+++ Tools/TestWebKitAPI/WebTransportServer.mm
@@ -29,10 +29,24 @@
 #import "HTTPServer.h"
 #import "Utilities.h"
 #import <pal/spi/cf/CFNetworkSPI.h>
+#import <wtf/BlockPtr.h>
 
 namespace TestWebKitAPI {
 
-WebTransportServer::WebTransportServer()
+struct WebTransportServer::Data : public RefCounted<WebTransportServer::Data> {
+    static Ref<Data> create(Function<Task(Connection)>&& connectionHandler) { return adoptRef(*new Data(WTFMove(connectionHandler))); }
+    Data(Function<Task(Connection)>&& connectionHandler)
+        : connectionHandler(WTFMove(connectionHandler)) { }
+
+    Function<Task(Connection)> connectionHandler;
+    RetainPtr<nw_listener_t> listener;
+    Vector<RetainPtr<nw_connection_group_t>> connectionGroups;
+    Vector<RetainPtr<nw_connection_t>> connections;
+    Vector<CoroutineHandle<Task::promise_type>> coroutineHandles;
+};
+
+WebTransportServer::WebTransportServer(Function<Task(Connection)>&& connectionHandler)
+    : m_data(Data::create(WTFMove(connectionHandler)))
 {
     auto configureTLS = [](nw_protocol_options_t options) {
         RetainPtr securityOptions = adoptNS(nw_quic_connection_copy_sec_protocol_options(options));
@@ -42,55 +56,45 @@ WebTransportServer::WebTransportServer()
 
     RetainPtr parameters = adoptNS(nw_parameters_create_quic_stream(NW_PARAMETERS_DEFAULT_CONFIGURATION, configureTLS));
 
-    m_listener = adoptNS(nw_listener_create(parameters.get()));
+    RetainPtr listener = adoptNS(nw_listener_create(parameters.get()));
 
-    // FIXME: This block unsafely captures this.
-    nw_listener_set_new_connection_group_handler(m_listener.get(), ^(nw_connection_group_t connectionGroup) {
+    nw_listener_set_new_connection_group_handler(listener.get(), [data = m_data] (nw_connection_group_t connectionGroup) {
         constexpr uint32_t maximumMessageSize { std::numeric_limits<uint32_t>::max() };
         constexpr bool rejectOversizedMessages { false };
-        nw_connection_group_set_receive_handler(connectionGroup, maximumMessageSize, rejectOversizedMessages, ^(dispatch_data_t data, nw_content_context_t, bool) {
+        nw_connection_group_set_receive_handler(connectionGroup, maximumMessageSize, rejectOversizedMessages, ^(dispatch_data_t, nw_content_context_t, bool) {
             // FIXME: Implement and test datagrams with WebTransport.
         });
 
-        // FIXME: This block unsafely captures this.
-        nw_connection_group_set_new_connection_handler(connectionGroup, ^(nw_connection_t connection) {
+        nw_connection_group_set_new_connection_handler(connectionGroup, [data] (nw_connection_t connection) {
+            data->connections.append(connection);
             nw_connection_set_queue(connection, dispatch_get_main_queue());
-            nw_connection_set_state_changed_handler(connection, ^(nw_connection_state_t state, nw_error_t error) {
-                if (error) {
-                    ASSERT_NOT_REACHED();
-                    return;
-                }
-                if (state != nw_connection_state_ready)
-                    return;
-                nw_connection_receive(connection, 1, std::numeric_limits<uint32_t>::max(), ^(dispatch_data_t content, nw_content_context_t context, bool is_complete, nw_error_t error) {
-                    nw_connection_send(connection, content, NW_CONNECTION_DEFAULT_MESSAGE_CONTEXT, true, ^(nw_error_t error) {
-                        ASSERT_UNUSED(error, !error);
-                    });
-                });
-            });
             nw_connection_start(connection);
-            m_connections.append(connection);
+            data->coroutineHandles.append(data->connectionHandler(connection).handle);
         });
 
         nw_connection_group_set_queue(connectionGroup, dispatch_get_main_queue());
         nw_connection_group_start(connectionGroup);
-        m_connectionGroups.append(connectionGroup);
+        data->connectionGroups.append(connectionGroup);
     });
-    nw_listener_set_queue(m_listener.get(), dispatch_get_main_queue());
+    nw_listener_set_queue(listener.get(), dispatch_get_main_queue());
 
     __block bool ready = false;
-    nw_listener_set_state_changed_handler(m_listener.get(), ^(nw_listener_state_t state, nw_error_t error) {
+    nw_listener_set_state_changed_handler(listener.get(), ^(nw_listener_state_t state, nw_error_t error) {
         ASSERT_UNUSED(error, !error);
         if (state == nw_listener_state_ready)
             ready = true;
     });
-    nw_listener_start(m_listener.get());
+    nw_listener_start(listener.get());
     Util::run(&ready);
+
+    m_data->listener = WTFMove(listener);
 }
+
+WebTransportServer::~WebTransportServer() = default;
 
 uint16_t WebTransportServer::port() const
 {
-    return nw_listener_get_port(m_listener.get());
+    return nw_listener_get_port(m_data->listener.get());
 }
 
 } // namespace TestWebKitAPI
