+++ Tools/TestWebKitAPI/cocoa/HTTPServer.mm
@@ -33,7 +33,6 @@
 #import <wtf/CallbackAggregator.h>
 #import <wtf/CompletionHandler.h>
 #import <wtf/RetainPtr.h>
-#import <wtf/SHA1.h>
 #import <wtf/ThreadSafeRefCounted.h>
 #import <wtf/text/Base64.h>
 #import <wtf/text/MakeString.h>
@@ -326,36 +325,6 @@ static ASCIILiteral statusText(unsigned statusCode)
     return "Unknown Status Code"_s;
 }
 
-static RetainPtr<dispatch_data_t> dataFromString(String&& s)
-{
-    auto impl = s.releaseImpl();
-    ASSERT(impl->is8Bit());
-    auto characters = impl->span8();
-    return adoptNS(dispatch_data_create(characters.data(), characters.size(), dispatch_get_main_queue(), ^{
-        (void)impl;
-    }));
-}
-
-RetainPtr<dispatch_data_t> dataFromVector(Vector<uint8_t>&& v)
-{
-    auto bufferSize = v.size();
-    auto rawPointer = v.releaseBuffer().leakPtr();
-    return adoptNS(dispatch_data_create(rawPointer, bufferSize, dispatch_get_main_queue(), ^{
-        fastFree(rawPointer);
-    }));
-}
-
-static Vector<uint8_t> vectorFromData(dispatch_data_t content)
-{
-    ASSERT(content);
-    __block Vector<uint8_t> request;
-    dispatch_data_apply(content, ^bool(dispatch_data_t, size_t, const void* buffer, size_t size) {
-        request.append(std::span { static_cast<const uint8_t*>(buffer), size });
-        return true;
-    });
-    return request;
-}
-
 static void appendUTF8ToVector(Vector<uint8_t>& vector, const String& string)
 {
     vector.append(string.utf8().span());
@@ -458,134 +427,6 @@ WKWebViewConfiguration *HTTPServer::httpsProxyConfiguration() const
     auto viewConfiguration = adoptNS([WKWebViewConfiguration new]);
     [viewConfiguration setWebsiteDataStore:adoptNS([[WKWebsiteDataStore alloc] _initWithConfiguration:storeConfiguration.get()]).get()];
     return viewConfiguration.autorelease();
-}
-
-void Connection::receiveBytes(CompletionHandler<void(Vector<uint8_t>&&)>&& completionHandler, size_t minimumSize) const
-{
-    nw_connection_receive(m_connection.get(), minimumSize, std::numeric_limits<uint32_t>::max(), makeBlockPtr([connection = *this, completionHandler = WTFMove(completionHandler)](dispatch_data_t content, nw_content_context_t, bool, nw_error_t error) mutable {
-        if (error || !content)
-            return completionHandler({ });
-        completionHandler(vectorFromData(content));
-    }).get());
-}
-
-void Connection::receiveHTTPRequest(CompletionHandler<void(Vector<char>&&)>&& completionHandler, Vector<char>&& buffer) const
-{
-    receiveBytes([connection = *this, completionHandler = WTFMove(completionHandler), buffer = WTFMove(buffer)](Vector<uint8_t>&& bytes) mutable {
-        buffer.appendVector(WTFMove(bytes));
-        if (auto* doubleNewline = strnstr(buffer.data(), "\r\n\r\n", buffer.size())) {
-            if (auto* contentLengthBegin = strnstr(buffer.data(), "Content-Length", buffer.size())) {
-                size_t contentLength = atoi(contentLengthBegin + strlen("Content-Length: "));
-                size_t headerLength = doubleNewline - buffer.data() + strlen("\r\n\r\n");
-                if (buffer.size() - headerLength < contentLength)
-                    return connection.receiveHTTPRequest(WTFMove(completionHandler), WTFMove(buffer));
-            }
-            completionHandler(WTFMove(buffer));
-        } else
-            connection.receiveHTTPRequest(WTFMove(completionHandler), WTFMove(buffer));
-    });
-}
-
-ReceiveOperation Connection::awaitableReceiveHTTPRequest() const
-{
-    return { *this };
-}
-
-void ReceiveOperation::await_suspend(std::coroutine_handle<> handle)
-{
-    m_connection.receiveHTTPRequest([this, handle](Vector<char>&& result) mutable {
-        m_result = WTFMove(result);
-        handle();
-    });
-}
-
-void SendOperation::await_suspend(std::coroutine_handle<> handle)
-{
-    m_connection.send(WTFMove(m_data), [handle] (bool) mutable {
-        handle();
-    });
-}
-
-SendOperation Connection::awaitableSend(Vector<uint8_t>&& message)
-{
-    return { dataFromVector(WTFMove(message)), *this };
-}
-
-SendOperation Connection::awaitableSend(String&& message)
-{
-    return { dataFromString(WTFMove(message)), *this };
-}
-
-SendOperation Connection::awaitableSend(RetainPtr<dispatch_data_t>&& data)
-{
-    return { WTFMove(data), *this };
-}
-
-void Connection::send(String&& message, CompletionHandler<void()>&& completionHandler) const
-{
-    send(dataFromString(WTFMove(message)), [completionHandler = WTFMove(completionHandler)] (bool) mutable {
-        if (completionHandler)
-            completionHandler();
-    });
-}
-
-void Connection::send(Vector<uint8_t>&& message, CompletionHandler<void()>&& completionHandler) const
-{
-    send(dataFromVector(WTFMove(message)), [completionHandler = WTFMove(completionHandler)] (bool) mutable {
-        if (completionHandler)
-            completionHandler();
-    });
-}
-
-void Connection::sendAndReportError(Vector<uint8_t>&& message, CompletionHandler<void(bool)>&& completionHandler) const
-{
-    send(dataFromVector(WTFMove(message)), WTFMove(completionHandler));
-}
-
-void Connection::send(RetainPtr<dispatch_data_t>&& message, CompletionHandler<void(bool)>&& completionHandler) const
-{
-    nw_connection_send(m_connection.get(), message.get(), NW_CONNECTION_DEFAULT_MESSAGE_CONTEXT, true, makeBlockPtr([completionHandler = WTFMove(completionHandler)](nw_error_t error) mutable {
-        if (completionHandler)
-            completionHandler(!!error);
-    }).get());
-}
-
-void Connection::webSocketHandshake(CompletionHandler<void()>&& connectionHandler)
-{
-    receiveHTTPRequest([connection = Connection(*this), connectionHandler = WTFMove(connectionHandler)] (Vector<char>&& request) mutable {
-
-        auto webSocketAcceptValue = [] (const Vector<char>& request) {
-            constexpr auto* keyHeaderField = "Sec-WebSocket-Key: ";
-            const char* keyBegin = strnstr(request.data(), keyHeaderField, request.size()) + strlen(keyHeaderField);
-            ASSERT(keyBegin);
-            const char* keyEnd = strnstr(keyBegin, "\r\n", request.size() + (keyBegin - request.data()));
-            ASSERT(keyEnd);
-
-            const auto webSocketKeyGUID = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11"_span;
-            SHA1 sha1;
-            sha1.addBytes(byteCast<uint8_t>(std::span { keyBegin, keyEnd }));
-            sha1.addBytes(webSocketKeyGUID);
-            SHA1::Digest hash;
-            sha1.computeHash(hash);
-            return base64EncodeToString(hash);
-        };
-
-        connection.send(HTTPResponse(101, {
-            { "Upgrade"_s, "websocket"_s },
-            { "Connection"_s, "Upgrade"_s },
-            { "Sec-WebSocket-Accept"_s, webSocketAcceptValue(request) }
-        }).serialize(HTTPResponse::IncludeContentLength::No), WTFMove(connectionHandler));
-    });
-}
-
-void Connection::terminate(CompletionHandler<void()>&& completionHandler)
-{
-    nw_connection_set_state_changed_handler(m_connection.get(), makeBlockPtr([completionHandler = WTFMove(completionHandler)] (nw_connection_state_t state, nw_error_t error) mutable {
-        ASSERT_UNUSED(error, !error);
-        if (state == nw_connection_state_cancelled && completionHandler)
-            completionHandler();
-    }).get());
-    nw_connection_cancel(m_connection.get());
 }
 
 Vector<uint8_t> HTTPResponse::bodyFromString(const String& string)
